pyproject.toml
[tool.poetry]
name = "visualisation-app"
version = "0.1.0"
description = "App for visalisation CNN"
authors = ["Кирилл Гринев <korsnaike@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
bottle = "0.12.25"
bottle-websocket = "0.2.9"
cffi = "1.16.0"
contourpy = "1.2.1"
cycler = "0.12.1"
eel = "0.16.0"
filelock = "3.9.0"
fonttools = "4.51.0"
fsspec = "2023.4.0"
future = "1.0.0"
gevent = "24.2.1"
gevent-websocket = "0.10.1"
greenlet = "3.0.3"
intel-openmp = "2021.4.0"
jinja2 = "3.1.2"
kiwisolver = "1.4.5"
markupsafe = "2.1.3"
matplotlib = "3.8.4"
mkl = "2021.4.0"
mpmath = "1.3.0"
networkx = "3.2.1"
numpy = "1.26.3"
opencv-contrib-python = "4.9.0.80"
opencv-python = "4.9.0.80"
packaging = "24.0"
pillow = "10.2.0"
pycparser = "2.22"
pyparsing = "3.1.2"
python-dateutil = "2.9.0.post0"
six = "1.16.0"
sympy = "1.12"
tbb = "2021.11.0"
typing-extensions = "4.8.0"
whichcraft = "0.6.1"
zope-event = "5.0"
zope-interface = "6.3"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[tool.poe.tasks]
force-cuda121 = "python -m pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
